requirements.txt
flask
google-generativeai
python-dotenv
numpy
requests
